src/requirements.txt
# SLAP - Scoreboard Live Automation Platform
# Python dependencies

# Serial communication
pyserial>=3.5

# Web framework
flask>=2.3.0
flask-socketio>=5.3.0

# WebSocket support
python-socketio>=5.8.0
gevent>=23.0.0
gevent-websocket>=0.10.1

# Development/testing
pytest>=7.0.0
